zabbix_django/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
flake8==3.7.9
freeze==1.0.10
idna==2.9
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2019.3
pyzabbix==0.7.5
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
wincertstore==0.2
yapf==0.29.0
